+++ vectra/def/__init__.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 """
-Created on Thu Jan 25 14:10:35 2018
+Created on Thu Mar 08 11:13:51 2018
 
 @author: ldh
 """
 
-# test.py
+# __init__.py